yada-29 try to make use of java 9 modules - the beginning
