+++ auth/src/main/java/bonux/yada/auth/Roles.java
@@ -1,0 +1,5 @@
+package bonux.yada.auth;
+
+public enum Roles {
+    USER, ADMIN
+}
