+++ auth/src/main/java/module-info.java
@@ -1,0 +1,11 @@
+module bonux.yada.auth {
+    exports bonux.yada.auth;
+
+    requires spring.security.core;
+    requires spring.security.config;
+    requires spring.web;
+    requires spring.context;
+    requires spring.beans;
+    requires java.sql;
+    requires spring.jdbc;
+}
