--- src/main/java/bonux/yada/domain/DomainTodoBuilder.java
@@ -1,15 +1,0 @@
-package bonux.yada.domain;
-
-public final class DomainTodoBuilder {
-
-    public static Todo fromModel(bonux.yada.repos.model.Todo todo) {
-        return new Todo(
-                todo.id,
-                todo.task,
-                todo.taskState,
-                todo.closeReason,
-                todo.taskStart,
-                todo.taskEnd,
-                todo.version);
-    }
-}
