--- src/main/java/bonux/yada/security/Roles.java
@@ -1,5 +1,0 @@
-package bonux.yada.security;
-
-public enum Roles {
-    USER, ADMIN
-}
