--- src/main/java/bonux/yada/types/CloseReason.java
@@ -1,5 +1,0 @@
-package bonux.yada.types;
-
-public enum CloseReason {
-    DONE, ABANDONED
-}
