--- src/main/java/bonux/yada/types/TaskState.java
@@ -1,5 +1,0 @@
-package bonux.yada.types;
-
-public enum TaskState {
-    NEW, OPEN, CLOSED
-}
