--- src/test/java/bonux/yada/YadaSpringContextIntegrationTest.java
@@ -1,16 +1,0 @@
-package bonux.yada;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@DataJpaTest
-public class YadaSpringContextIntegrationTest {
-
-    @Test
-    public void testSetup() {
-
-    }
-}
