+++ todo/src/main/java/bonux/yada/todo/types/CloseReason.java
@@ -1,0 +1,5 @@
+package bonux.yada.todo.types;
+
+public enum CloseReason {
+    DONE, ABANDONED
+}
