+++ todo/src/main/java/bonux/yada/todo/types/TaskState.java
@@ -1,0 +1,5 @@
+package bonux.yada.todo.types;
+
+public enum TaskState {
+    NEW, OPEN, CLOSED
+}
